+++ frontend/src/charts/DisparityBarChart.tsx
@@ -7,11 +7,9 @@ import {
   BREAKDOWN_VAR_DISPLAY_NAMES,
 } from "../data/query/Breakdowns";
 import { MetricConfig } from "../data/config/MetricConfig";
-import {
-  addLineBreakDelimitersToField,
-  MULTILINE_LABEL,
-  AXIS_LABEL_Y_DELTA,
-} from "./utils";
+
+const DELIMITER = "*~*";
+const MAX_LINE_LENGTH = 18;
 
 function getSpec(
   data: Record<string, any>[],
@@ -31,6 +29,9 @@ function getSpec(
   const THICK_MEASURE_COLOR = "#BDC1C6";
   const DATASET = "DATASET";
   const WIDTH_PADDING_FOR_SNOWMAN_MENU = 50;
+  const MULTILINE_LABEL = `split(datum.value, '${DELIMITER}')`;
+  // We use nested ternerys to determine the label's y axis delta based on the number of lines in the label to vertically align
+  const AXIS_LABEL_Y_DELTA = `length(${MULTILINE_LABEL}) == 2 ? -3 : length(${MULTILINE_LABEL}) > 2 ? -7 : 5`;
 
   function maxValueInField(field: string) {
     return Math.max(
@@ -233,10 +234,20 @@ export function DisparityBarChart(props: DisparityBarChartProps) {
     100 /* default width during intialization */
   );
 
-  let dataWithLineBreakDelimiter = addLineBreakDelimitersToField(
-    props.data,
-    props.breakdownVar
-  );
+  let dataWithLineBreakDelimiter = props.data.map((data) => {
+    let lines = [];
+    let currentLine = "";
+    for (let word of data[props.breakdownVar].split(" ")) {
+      if (word.length + currentLine.length >= MAX_LINE_LENGTH) {
+        lines.push(currentLine.trim());
+        currentLine = word + " ";
+      } else {
+        currentLine += word + " ";
+      }
+    }
+    lines.push(currentLine.trim());
+    return { ...data, ...{ [props.breakdownVar]: lines.join(DELIMITER) } };
+  });
   dataWithLineBreakDelimiter.sort((a, b) =>
     a[props.breakdownVar].localeCompare(b[props.breakdownVar])
   );
